moved input und signals utils dir and put syntax checks to lexer and parser
