+++ include/lexer.h
@@ -38,5 +38,6 @@ void	set_type(t_token *new);
 void	del_first(t_list **lst);
 void	print_lexed_lst(t_list *lst);
 void	quotation_error(t_list *lst, char *str);
+int		check_token_syntax(t_list *token_list);
 
 #endif
