+++ include/parser.h
@@ -34,5 +34,6 @@ void			default_in_out(int *in, int *out, struct s_state *state);
 void			update_state(t_token *token, struct s_state *state, char *cmd);
 void			update_in_out_state(struct s_state *state, t_type type);
 int				open_files(char *path, int flag);
+int				check_parsed_syntax(t_list *command_table);
 
 #endif
