+++ src/parser/state_utils.c
@@ -71,31 +71,19 @@ void	update_state(t_token *token, struct s_state *state, char *cmd)
 	}
 }
 
-void	print_parsed_lst(t_list *command_table)
+int	check_parsed_syntax(t_list *command_table)
 {
-	t_command	*temp;
-	int			i;
+	t_command	*temp_c;
 
-	i = 0;
 	while (command_table != NULL)
 	{
-		temp = (t_command *)command_table->content;
-		ft_printf("------------------\n");
-		if (!temp->command)
-			return ;
-		ft_printf("Command: %s\n", temp->command);
-		ft_printf("Options: ");
-		while (temp && temp->options != NULL && temp->options[i] != NULL)
+		temp_c = (t_command *) command_table->content;
+		if (!temp_c)
 		{
-			ft_printf("%d: ", i);
-			ft_printf("%s ", temp->options[i]);
-			i++;
+			ft_putstr_fd("Syntax error, missing a command\n", STDERR_FILENO);
+			return (EXIT_FAILURE);
 		}
-		ft_printf("\n");
-		ft_printf("In: %d\n", temp->in);
-		ft_printf("Out: %d\n", temp->out);
-		ft_printf("------------------\n");
 		command_table = command_table->next;
-		i = 0;
 	}
+	return (EXIT_SUCCESS);
 }
